Last changes

+++ Project/Individual-Track-Project/demo/src/main/java/com/example/demo/Repositories/OrderItemRepository.java
@@ -1,9 +1,9 @@
 package com.example.demo.Repositories;
 
-import com.example.demo.Normal.Order;
+import com.example.demo.RepositoryClasses.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemRepository extends JpaRepository<Order, Integer> {
+public interface OrderItemRepository extends JpaRepository<Order, Long> {
 
     boolean existsById(Integer id);
 }
