+++ Project/Individual-Track-Project/demo/src/main/java/com/example/demo/dto/ClientDTO.java
@@ -1,9 +1,7 @@
 package com.example.demo.dto;
 
-import com.example.demo.Normal.User;
-
 public class ClientDTO extends UserDTO {
     public ClientDTO(Long id, String firstName, String lastName, String username, String password, String email, String address, String phone, boolean logged, String category) {
-        super(id, firstName, lastName, username, password, email, address, phone, logged, category);
+        super(Math.toIntExact(id), firstName, lastName, username, password, email, address, phone, logged, category);
     }
 }
